subscription added with multi tenant

+++ app/Http/Controllers/Backend/CarModelController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\Models\CarModel;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 class CarModelController extends Controller
@@ -24,7 +25,13 @@ class CarModelController extends Controller
 
     public function index()
     {
-        $carModels = CarModel::get();
+        $tenant = Auth::user()->currentTenant();
+
+        if (!$tenant) {
+            return redirect()->route('dashboard')
+                ->with('error', 'No active company found! Please contact administrator.');
+        }
+        $carModels = CarModel::where('tenant_id', $tenant->id)->get();
         return view($this->dir. 'index', compact('carModels'));
     }
 
@@ -35,10 +42,13 @@ class CarModelController extends Controller
 
     public function store(Request $request)
     {
+        $tenant = Auth::user()->currentTenant();
+
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:car_models',
         ]);
-
+        $validated['tenant_id'] = $tenant->id;
+        $validated['createdBy'] = Auth::id();
         CarModel::create($validated);
 
         return redirect()->route('car-models.index')
@@ -47,16 +57,26 @@ class CarModelController extends Controller
 
     public function edit($id)
     {
-        $model = CarModel::findOrFail($id);
+        $tenant = Auth::user()->currentTenant();
+
+        if (!$tenant) {
+            return redirect()->route('dashboard')
+                ->with('error', 'No active company found!');
+        }
+
+        $model = CarModel::where('tenant_id', $tenant->id)->findOrFail($id);
         return view($this->dir.'edit', compact('model'));
     }
 
     public function update(Request $request, CarModel $carModel)
     {
+        $tenant = Auth::user()->currentTenant();
+
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:car_models,name,' . $carModel->id,
         ]);
-
+        $validated['tenant_id'] = $tenant->id;
+        $validated['updatedBy'] = Auth::id();
         $carModel->update($validated);
 
         return redirect()->route('car-models.index')
@@ -65,6 +85,12 @@ class CarModelController extends Controller
 
     public function destroy(CarModel $carModel)
     {
+        $tenant = Auth::user()->currentTenant();
+
+        // ✅ Check ownership
+        if ($carModel->tenant_id !== $tenant->id) {
+            abort(403, 'Unauthorized access');
+        }
         $carModel->delete();
 
         return redirect()->route('car-models.index')
